refactor(error): refactor error type file name
